requirements.txt
opencv-python
numpy
Pillow
imagehash
faiss-cpu
tqdm
matplotlib
piexif
